App/RoomsListViewItem - + extracted RoomsListViewItem to it's own file.
